requirements.txt
# Core dependencies
fastapi==0.110.0
python-telegram-bot[job-queue]==21.4
uvicorn==0.27.1

# Database
asyncpg==0.29.0
psycopg[binary]==3.2.3
psycopg_pool==3.2.3

# Google Sheets integration
gspread==6.1.2
google-auth==2.34.0

# Utils
python-dotenv==1.0.1
pandas==2.3.3
cachetools==5.3.3
APScheduler==3.10.4